refactoring - move health monitor feature from NioServer to HealthMonitor
